+++ plugin/resources/assets/js/api/Plagiarism.js
@@ -34,6 +34,36 @@ class Plagiarism {
                 VueEvent.$emit('show-notification', 'Error fetching plagiarism matches.\n' + error, 'danger')
             })
     }
+
+    static runPlagiarismCheck(courseId, charonId, then) {
+        axios.post(`${this.getRoot()}/courses/${courseId}/charons/${charonId}/plagiarism/run`)
+            .then(response => {
+                then(response)
+            })
+            .catch(error => {
+                VueEvent.$emit('show-notification', 'Error running plagiarism check.\n' + error, 'danger')
+            })
+    }
+
+    static getLatestCheckStatus(charonId, checkId, then) {
+        axios.get(`${this.getRoot()}/charons/${charonId}/plagiarism-checks/${checkId}`)
+            .then(response => {
+                then(response.data)
+            })
+            .catch(error => {
+                VueEvent.$emit('show-notification', 'Error getting status of latest plagiarism check.\n' + error, 'danger')
+            })
+    }
+
+    static getCheckHistory(courseId, then) {
+        axios.get(`${this.getRoot()}/courses/${courseId}/checks-history/`)
+            .then(response => {
+                then(response.data)
+            })
+            .catch(error => {
+                VueEvent.$emit('show-notification', 'Error getting checks history.\n' + error, 'danger')
+            })
+    }
 }
 
 export default Plagiarism